refactor: split ssl-functionality of MocketSocket into MocketSSLSocket
