+++ mocket/ssl/context.py
@@ -7,6 +7,7 @@ from typing import Any
 import urllib3.util.ssl_
 
 from mocket.socket import MocketSocket
+from mocket.ssl.socket import MocketSSLSocket
 
 true_ssl_context = ssl.SSLContext
 
@@ -70,10 +71,26 @@ class FakeSSLContext(SuperFakeSSLContext):
             setattr(self, m, dummy_method)
 
     @staticmethod
-    def wrap_socket(sock: MocketSocket, *args: Any, **kwargs: Any) -> MocketSocket:
-        sock._kwargs = kwargs
-        sock._secure_socket = True
-        return sock
+    def wrap_socket(sock: MocketSocket, *args: Any, **kwargs: Any) -> MocketSSLSocket:
+        ssl_socket = MocketSSLSocket()
+        ssl_socket._original_socket = sock
+
+        ssl_socket._true_socket = true_urllib3_ssl_wrap_socket(
+            sock._true_socket,
+            **kwargs,
+        )
+        ssl_socket._kwargs = kwargs
+
+        ssl_socket._timeout = sock._timeout
+
+        ssl_socket._host = sock._host
+        ssl_socket._port = sock._port
+        ssl_socket._address = sock._address
+
+        ssl_socket._io = sock._io
+        ssl_socket._entry = sock._entry
+
+        return ssl_socket
 
     @staticmethod
     def wrap_bio(
@@ -81,7 +98,7 @@ class FakeSSLContext(SuperFakeSSLContext):
         outgoing: Any,  # _ssl.MemoryBIO
         server_side: bool = False,
         server_hostname: str | bytes | None = None,
-    ) -> MocketSocket:
-        ssl_obj = MocketSocket()
+    ) -> MocketSSLSocket:
+        ssl_obj = MocketSSLSocket()
         ssl_obj._host = server_hostname
         return ssl_obj
